+++ kitty/launch.py
@@ -185,6 +185,37 @@ active window. The default is to place the window in a layout dependent manner,
 typically, after the currently active window.
 
 
+--bias
+type=float
+default=0
+The bias used to alter the size of the window.
+It controls what fraction of available space the window takes. The exact meaning
+of bias depends on the current layout.
+
+* Splits layout: The bias is interpreted as a percentage between 0 and 100.
+When splitting a window into two, the new window will take up the specified fraction
+of the space alloted to the original window and the original window will take up
+the remainder of the space.
+
+* Vertical/horizontal layout: The bias is interpreted as adding/subtracting from the
+normal size of the window. It should be a number between -90 and 90. This number is
+the percentage of cells in the full OS window that should be added to the window size.
+So for example, if a window would normally have been 50 cells in the layout inside an
+OS Window that is 80 cells high and --bias -10 is used it will become *approximately*
+42 cells high. Note that cell counts are approximations, you cannot use this method to
+create windows of fixed cell sizes.
+
+* Tall layout: If the window being created is the *first* window in a column, then
+the bias is interpreted as a percentage, as for the splits layout, splitting the OS
+Window width between columns. If the window is a second or subsequent window in a column
+the bias is interpreted as adding/subtracting from the window size as for the vertical
+layout above.
+
+* Fat layout: Same as tall layout except it goes by rows instead of columns.
+
+The bias option was introduced in kitty version 0.36.0.
+
+
 --allow-remote-control
 type=bool-set
 Programs running in this window can control kitty (even if remote control is not
@@ -433,6 +464,7 @@ class LaunchKwds(TypedDict):
     overlay_for: Optional[int]
     stdin: Optional[bytes]
     hold: bool
+    bias: Optional[float]
 
 
 def apply_colors(window: Window, spec: Sequence[str]) -> None:
@@ -520,11 +552,14 @@ def _launch(
         'overlay_for': None,
         'stdin': None,
         'hold': False,
+        'bias': None,
     }
     spacing = {}
     if opts.spacing:
         from .rc.set_spacing import parse_spacing_settings, patch_window_edges
         spacing = parse_spacing_settings(opts.spacing)
+    if opts.bias:
+        kw['bias'] = max(-100, min(opts.bias, 100))
     if opts.cwd:
         if opts.cwd == 'current':
             if active: