+++ kitty/rc/launch.py
@@ -50,6 +50,7 @@ class Launch(RemoteCommand):
     os_window_state/choices.normal.fullscreen.maximized.minimized: The initial state for OS Window
     color/list.str: list of color specifications such as foreground=red
     watcher/list.str: list of paths to watcher files
+    bias/float: The bias with which to create the new window in the current layout
     '''
 
     short_desc = 'Run an arbitrary process in a new window/tab'